// src/types.ts
export type FastlyMetadata = {
    client: object | null
    'x-sigsci-tags-header': string | null
    'x-sigsci-requestid-header': string | null
}

export type Env = {
    ORIGIN_URL: string
    SURFACE_SLUG: string
    AUTHENTICATED_USER_JWT_COOKIE_NAME: string
    ANONYMOUS_SESSION_COOKIE_NAME: string
    INJECT_SCRIPT_URL: string
    MONETIZATION_OS_HOST: string
    MONETIZATION_OS_ENDPOINTS_PREFIX: string
    MONETIZATION_OS_SECRET_KEY: string
    SURFACE_DECISIONS_IGNORE_PATHS?: string
}

export type PageMetadata = Record<string, string>

export interface FeatureMeterableProperty {
    type: 'meterable'
    hasAccess: boolean
    remainingUnits?: number
}

export interface FeatureNumberProperty {
    type: 'number'
    value: number
}

export interface Feature {
    featureSlug: string
    properties: Record<string, FeatureMeterableProperty | FeatureNumberProperty>
    sideEffects: unknown[]
}

export interface SubSurfaceMetadataApi {
    cssSelector?: string | null
}

export type SetHttpResponse = {
    headers?: Record<string, string>
    cookies?: string[]
    status: number
    statusText?: string
    body: string | null
}

export type ModifyHttpResponse = {
    addHeaders?: { name: string; value: string }[]
    removeHeaders?: string[]
    addCookies?: string[]
    status?: number
    statusText?: string
    body?: string | null
}

export type SurfaceBehaviorApi = {
    http?: ModifyHttpResponse | SetHttpResponse
    properties?: Record<string, unknown>
} & Record<string, unknown>

export type WebComponentElement<T = Record<string, unknown>> = {
    schema: string
    props: T
}

export type WebElement =
    | {
          type: 'html'
          content: string
      }
    | {
          type: 'text'
          content: string
      }
    | ({
          type: 'element'
      } & WebComponentElement)
    | ({
          type: 'custom'
      } & Record<string, unknown>)

export type WebComponentRangeReplacement = {
    fromMarker?: string
    toMarker?: string
    replaceWith?: WebElement[] | null
}

export type WebContentSurfaceBehavior = {
    before?: WebElement[]
    prepend?: WebElement[]
    remove?: boolean
    replaceRange?: WebComponentRangeReplacement | null
    append?: WebElement[]
    after?: WebElement[]
}

export type SubSurfaceBehaviorApi = {
    content?: WebContentSurfaceBehavior
    properties?: Record<string, unknown>
    metadata: SubSurfaceMetadataApi
} & Record<string, unknown>

export interface SurfaceDecisionResponse {
    status: 'success'
    identity: {
        identifier: string
        isAuthenticated: boolean
        authType: string
        jwtClaims: Record<string, unknown>
    }
    features: Record<string, Feature>
    customer: {
        hasProducts: boolean
    }
    surfaceBehavior: SurfaceBehaviorApi
    componentsSkipped: boolean
    componentBehaviors: Record<string, SubSurfaceBehaviorApi>
}

export interface SurfaceDecisionError {
    message: string
    status: 'error'
    statusCode: number
}